interrupts work (kinda)

+++ include/types.h
@@ -2,14 +2,14 @@
 #define _TYPES_H
 #define NULL 0
 
-typedef unsigned long  u64;
-typedef          long  s64;
-typedef unsigned int   u32;
-typedef          int   s32;
-typedef unsigned short u16;
-typedef          short s16;
-typedef unsigned char   u8;
-typedef          char   s8;
+typedef unsigned long long u64;
+typedef          long long s64;
+typedef unsigned int       u32;
+typedef          int       s32;
+typedef unsigned short     u16;
+typedef          short     s16;
+typedef unsigned char      u8;
+typedef          char      s8;
 
 typedef int bool;
 #define true 1
